DTSCCI-378 refactoring consumer tests
